requirements.txt
streamlit>=1.32
transformers>=4.40
torch>=2.2
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
requests>=2.31
beautifulsoup4>=4.12
newspaper3k>=0.2.8
lxml>=5.1
lxml_html_clean
python-dotenv>=1.0
apscheduler>=3.10
feedparser>=6.0
joblib>=1.3
tldextract>=5.1
plotly>=5.20
